requirements.txt
Flask
waitress
google-cloud-storage
